Big commit. Its easier to say what I didnt do than to say I did

+++ core/domain-model/src/main/java/ru/moonlight/domain_model/catalog/ProductMetadataDomainModel.kt
@@ -1,7 +1,7 @@
 package ru.moonlight.domain_model.catalog
 
 class ProductMetadataDomainModel(
-    val popularSizes: List<Double>,
-    val minPrice: Double,
-    val maxPrice: Double,
+    val popularSizes: List<Float>,
+    val minPrice: Float,
+    val maxPrice: Float,
 )
